+++ lua/fzfx/detail/popup.lua
@@ -1,6 +1,6 @@
+local constants = require("fzfx.lib.constants")
 local numbers = require("fzfx.commons.numbers")
 local apis = require("fzfx.commons.apis")
-local shells = require("fzfx.lib.shells")
 local fileios = require("fzfx.commons.fileios")
 local log = require("fzfx.lib.log")
 
@@ -38,6 +38,71 @@ function WindowOptsContext:restore()
 end
 
 -- WindowOptsContext }
+
+-- ShellOptsContext {
+
+--- @class fzfx.ShellOptsContext
+--- @field shell string?
+--- @field shellslash string?
+--- @field shellcmdflag string?
+--- @field shellxquote string?
+--- @field shellquote string?
+--- @field shellredir string?
+--- @field shellpipe string?
+--- @field shellxescape string?
+local ShellOptsContext = {}
+
+--- @return fzfx.ShellOptsContext
+function ShellOptsContext:save()
+  local o = constants.IS_WINDOWS
+      and {
+        shell = vim.o.shell,
+        shellslash = vim.o.shellslash,
+        shellcmdflag = vim.o.shellcmdflag,
+        shellxquote = vim.o.shellxquote,
+        shellquote = vim.o.shellquote,
+        shellredir = vim.o.shellredir,
+        shellpipe = vim.o.shellpipe,
+        shellxescape = vim.o.shellxescape,
+      }
+    or {
+      shell = vim.o.shell,
+    }
+  setmetatable(o, self)
+  self.__index = self
+
+  if constants.IS_WINDOWS then
+    vim.o.shell = "cmd.exe"
+    vim.o.shellslash = false
+    vim.o.shellcmdflag = "/s /c"
+    vim.o.shellxquote = '"'
+    vim.o.shellquote = ""
+    vim.o.shellredir = ">%s 2>&1"
+    vim.o.shellpipe = "2>&1| tee"
+    vim.o.shellxescape = ""
+  else
+    vim.o.shell = "sh"
+  end
+
+  return o
+end
+
+function ShellOptsContext:restore()
+  if constants.IS_WINDOWS then
+    vim.o.shell = self.shell
+    vim.o.shellslash = self.shellslash
+    vim.o.shellcmdflag = self.shellcmdflag
+    vim.o.shellxquote = self.shellxquote
+    vim.o.shellquote = self.shellquote
+    vim.o.shellredir = self.shellredir
+    vim.o.shellpipe = self.shellpipe
+    vim.o.shellxescape = self.shellxescape
+  else
+    vim.o.shell = self.shell
+  end
+end
+
+-- ShellOptsContext }
 
 --- @class fzfx.PopupWindowConfig
 --- @field anchor "NW"|nil
@@ -441,7 +506,7 @@ function Popup:new(win_opts, source, fzf_opts, actions, context, on_popup_exit)
   end
 
   -- save shell opts
-  local shell_opts_context = shells.ShellOptsContext:save()
+  local shell_opts_context = ShellOptsContext:save()
   local prev_fzf_default_opts = vim.env.FZF_DEFAULT_OPTS
   local prev_fzf_default_command = vim.env.FZF_DEFAULT_COMMAND
   vim.env.FZF_DEFAULT_OPTS = fzf_helpers.make_fzf_default_opts()
