+++ lua/fzfx/lib/shells.lua
@@ -2,69 +2,6 @@ local constants = require("fzfx.lib.constants")
 
 local M = {}
 
---- @class fzfx.ShellOptsContext
---- @field shell string?
---- @field shellslash string?
---- @field shellcmdflag string?
---- @field shellxquote string?
---- @field shellquote string?
---- @field shellredir string?
---- @field shellpipe string?
---- @field shellxescape string?
-local ShellOptsContext = {}
-
---- @return fzfx.ShellOptsContext
-function ShellOptsContext:save()
-  local o = constants.IS_WINDOWS
-      and {
-        shell = vim.o.shell,
-        shellslash = vim.o.shellslash,
-        shellcmdflag = vim.o.shellcmdflag,
-        shellxquote = vim.o.shellxquote,
-        shellquote = vim.o.shellquote,
-        shellredir = vim.o.shellredir,
-        shellpipe = vim.o.shellpipe,
-        shellxescape = vim.o.shellxescape,
-      }
-    or {
-      shell = vim.o.shell,
-    }
-  setmetatable(o, self)
-  self.__index = self
-
-  if constants.IS_WINDOWS then
-    vim.o.shell = "cmd.exe"
-    vim.o.shellslash = false
-    vim.o.shellcmdflag = "/s /c"
-    vim.o.shellxquote = '"'
-    vim.o.shellquote = ""
-    vim.o.shellredir = ">%s 2>&1"
-    vim.o.shellpipe = "2>&1| tee"
-    vim.o.shellxescape = ""
-  else
-    vim.o.shell = "sh"
-  end
-
-  return o
-end
-
-function ShellOptsContext:restore()
-  if constants.IS_WINDOWS then
-    vim.o.shell = self.shell
-    vim.o.shellslash = self.shellslash
-    vim.o.shellcmdflag = self.shellcmdflag
-    vim.o.shellxquote = self.shellxquote
-    vim.o.shellquote = self.shellquote
-    vim.o.shellredir = self.shellredir
-    vim.o.shellpipe = self.shellpipe
-    vim.o.shellxescape = self.shellxescape
-  else
-    vim.o.shell = self.shell
-  end
-end
-
-M.ShellOptsContext = ShellOptsContext
-
 --- @param s string
 --- @param special any?
 --- @return string
